refined file paths to get mockr working; moved engine.rb to lib/bushido/engine.rb and hooks.rb to lib/hooks.rb
