+++ lib/hooks.rb
